requirements.txt
anthropic
openai
python-dotenv
Pillow
rich
prompt_toolkit
pymongo==4.5.0
pyperclip==1.8.2